move kademlia related modules to separate namespace
