+++ apps/ex_wire/test/ex_wire/kademlia/node_test.exs
@@ -1,0 +1,4 @@
+defmodule ExWire.Kademlia.NodeTest do
+  use ExUnit.Case, async: true
+  doctest ExWire.Kademlia.Node
+end
