--- apps/ex_wire/test/ex_wire/struct/node_test.exs
@@ -1,4 +1,0 @@
-defmodule ExWire.Struct.NodeTest do
-  use ExUnit.Case, async: true
-  doctest ExWire.Struct.Node
-end
